backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
boto3==1.42.7
botocore==1.42.7
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0
django-storages==1.14.6
djangorestframework==3.16.1
exceptiongroup==1.3.1
idna==3.11
jmespath==1.0.1
kombu==5.6.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==6.4.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
vine==5.1.0
wcwidth==0.2.14

gunicorn